clean: separated files
