+++ src/algorithms/nsga_ii.rs
@@ -5,6 +5,8 @@
 //! file, You can obtain one at https://www.gnu.org/licenses/gpl-3.0.html
 
 mod algorithm;
+mod individual;
+mod core;
 
 use crate::graph::{Graph, Partition};
 use crate::utils::args::AGArgs;
@@ -12,7 +14,7 @@ use crate::utils::args::AGArgs;
 /// Main run function that creates both the real and random fronts, then
 /// selects the best solution via the chosen criterion.
 pub fn run(graph: &Graph, args: AGArgs) -> (Partition, Vec<f64>, f64) {
-    let partition = algorithm::detect_communities(graph, args.debug);
+    let partition = algorithm::nsga_ii(graph, args.debug, args);
 
     (partition, Vec::new(), 0.0)
 }
